+++ src/Properties/AssemblyInfo.cs
@@ -55,6 +55,6 @@ using System.Windows;
 // You can specify all the values or you can default the Build and Revision Numbers
 // by using the '*' as shown below:
 // [assembly: AssemblyVersion("1.0.*")]
-// Bumped to 1.1.1: fix taskbar icon, fix theme switching, sync installer version
-[assembly: AssemblyVersion("1.1.1.0")]
-[assembly: AssemblyFileVersion("1.1.1.0")]
+// Bumped to 1.1.2: fix silent crash on launch from malformed ICO in XAML Icon property
+[assembly: AssemblyVersion("1.1.2.0")]
+[assembly: AssemblyFileVersion("1.1.2.0")]